Fix CI Build Errors and Warnings
- Hardware.Control: Fixed openFd signature mismatch for unix-2.8.0.0 (CI) using CPP to maintain local compatibility.
- Numeric.Simple: Fixed partial function `head` in inverse.
- SignalProcessing.FMCW: Fixed partial functions `tail`/`init` in unwrapPhase. Added explicit type annotations to fix defaulting warnings.
- SignalProcessing.Regression: Fixed type defaulting warnings.
- SignalProcessing.Kalman: Removed unused bindings.
- Data.Time.HighRes: Fixed unused record fields warnings.
- app/Main.hs: Fixed export list warning.

+++ src/SignalProcessing/Regression.hs
@@ -12,7 +12,7 @@ solveBiQuadratic x y
     | length x /= length y = replicate 5 0.0
     | otherwise = leastSquares designM y
   where
-    designM = map (\val -> [1, val, val^2, val^3, val^4]) x
+    designM = map (\val -> [1, val, val^(2::Int), val^(3::Int), val^(4::Int)]) x
 
 -- | Perform the Regression for "Strict" Bi-Quadratic
 -- y = a*x^4 + b*x^2 + c
@@ -22,7 +22,7 @@ solveStrictBiQuadratic x y
     | length x /= length y = replicate 3 0.0
     | otherwise = leastSquares designM y
   where
-    designM = map (\val -> [1, val^2, val^4]) x
+    designM = map (\val -> [1, val^(2::Int), val^(4::Int)]) x
 
 -- | Prediction Function
 predict :: [Double] -> Double -> Double
